code/data/requirements.txt
geatpy==2.7.0
geopy==2.4.1
matplotlib==3.5.3
numpy==1.24.4
openpyxl==3.1.2
pandas @ file:///C:/miniconda3/conda-bld/pandas_1692299636855/work
pillow @ file:///C:/b/abs_e22m71t0cb/croot/pillow_1707233126420/work
pyarrow==15.0.2
scikit-learn==1.3.2
seaborn==0.13.2
tensorflow==2.10.0
